+++ src/tournaments-teams/dto/create-tournaments-team.dto.ts
@@ -1,9 +1,5 @@
-import { IsNumber } from "class-validator";
-
 export class CreateTournamentsTeamDto {
-    @IsNumber()
     tournamentId: number;
-    
     name: string;
     city: string;
 }
