+++ src/tournaments-teams/entities/tournaments-team.entity.ts
@@ -1,6 +1,16 @@
+import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+
+@Entity("tournament_teams")
 export class TournamentsTeam {
+    @Column({ name: 'tournament_id'})
     tournamentId: number;
+
+    @PrimaryGeneratedColumn()
     id: number;
+
+    @Column()
     name: string;
+
+    @Column()
     city: string;
 }
